RH-8: adding clasificadores from presupuestos

+++ src/modules/rrhh/RegistroPersonal/components/RegistroPersonal.vue
@@ -4,7 +4,7 @@
   <div class="table-headers">
     <div class="d-flex gap-1 mb-2">
       <CButton color="info" @click="showModal">Agregar</CButton>
-      <CButton color="secondary" @click="() => { reportes = true }">Imprimir Reporte</CButton>
+      <CButton color="secondary">Imprimir Reporte</CButton>
     </div>
   </div>
 
@@ -14,7 +14,7 @@
     :actions="buttonActions"
     :footer="footerItem"
   />
-  
+
   <RegistroPersonalDialog
     :showModal="showRegistroPersonalModal"
     @post-personal="submitForm"
@@ -38,7 +38,7 @@
     :showModal="showTipoNovedad"
     @closeModal="closeTipoNovedad"
   />
-  
+
   <ContenedorArchivosRRHH
     :showModal="showModalDoc"
     :empleado="selectedEmployee"
@@ -48,7 +48,7 @@
   <EducacionDialog
     :showModal="showEducacion"
     :employeeInfo="selectedEmployee"
-    @closeModal="() => showEducacion = false"
+    @closeModal="() => (showEducacion = false)"
   />
 </template>
 
@@ -68,6 +68,7 @@ import ContenedorArchivosRRHH from './ContenedorArchivosRRHH.vue'
 import AccionPersonalDialog from './Dialogos/AccionPersonalDialog.vue'
 import TipoNovedadDialog from './TipoNovedades.vue'
 import TarjetaEmpleadoDialogs from '../components/Dialogos/TarjetaEmpleado.vue'
+import Swal from 'sweetalert2/dist/sweetalert2.js'
 
 export default {
   components: {
@@ -86,155 +87,24 @@ export default {
 
   data: function () {
     return {
-      selectedEmployee:{},
-      filedata: {},
-      userid: null,
-      status: false,
+      selectedEmployee: {},
       showModalDoc: false,
-      accionEmpleado: {},
-      accionPersonal: [],
       showAccionPersonal: false,
       showTipoNovedad: false,
-      empleado: null,
       newTarjetaEmpleadoModal: false,
       showRegistroPersonalModal: false,
-      id: null,
-      empleadoReporte: {},
-      showTarjeta: false,
-      showModalRepots: false,
-      employeeInfo: {},
       showEducacion: false,
-      employeeInfo: null,
-      showRegistroPersonalModal: false,
-      id: null,
-      actions: [],
-      lgDemo4: false,
-      cambiar: false,
       registroPersonal: [],
-      horaActual: '',
-      toasts: [],
-      code: null,
-      noEnviado: false,
-      Error: '',
-      status: null,
-      liveExampleVisible: false,
-      reporteDepto: 1,
-      reportes: false,
-      posicionCargo: [{}],
-      tipoSangre: [{}],
-      areaTrabajo: [{}],
-      programaDivision: [{}],
-      sector: [{}],
-      departamentos: [],
-      postFile: {
-
-      },
-      postEmpleado: {
-        ayuntamientoId: null,
-        codigo: null,
-        nombres: null,
-        apellidos: null,
-        tipoDocumento: null,
-        cedula: null,
-        direccion: null,
-        sectorId: 1,
-        telefono: null,
-        celular: null,
-        fechaNacimiento: new Date(Date.now()),
-        lugarNacimiento: null,
-        estadoCivil: 'Soltero',
-        sexo: 'M',
-        dependientes: 0,
-        fechaIngreso: new Date(),
-        fechaSalida: new Date(Date.now()),
-        razonSalida: null,
-        reemplear: true,
-        fechaReingreso: new Date(Date.now()),
-        programaDivisionId: 0,
-        departamentoId: 0,
-        areaTrabajoId: 0,
-        posicionId: 0,
-        grupoOcupacional: null,
-        tipoContrato: null,
-        fechaInicioContrato: new Date(Date.now()),
-        fechaFinContrato: new Date(Date.now()),
-        turno: null,
-        periodoPago: null,
-        formaPago: null,
-        numeroCuenta: null,
-        fechaExpitaTarjeta: new Date(Date.now()),
-        estatus: true,
-        sueldo: 0.0,
-        sueldoAnterior: 0.0,
-        fechaSueldoAnterior: new Date(Date.now()),
-        fechaUltimaNomina: new Date(Date.now()),
-        inicioVacaciones: new Date(Date.now()),
-        finVacaciones: new Date(Date.now()),
-        activoNomina: true,
-        ingreso2: 0.0,
-        ingreso3: 0.0,
-        ingreso4: 0.0,
-        ingreso5: 0.0,
-        ingreso6: 0.0,
-        ingreso7: 0.0,
-        ingreso8: 0.0,
-        ingreso9: 0.0,
-        ingreso10: 0.0,
-        impuestoSobreRenta: 0.0,
-        arsCalculado: true,
-        arsFijo: 0.0,
-        afpCalculado: true,
-        afpFijo: 0.0,
-        egresos4: 0.0,
-        egresos5: 0.0,
-        egresos6: 0.0,
-        egresos7: 0.0,
-        egresos8: 0.0,
-        egresos9: 0.0,
-        egresos10: 0.0,
-        eneroIngreso: 0.0,
-        febreroIngreso: 0.0,
-        marzoIngreso: 0.0,
-        abrilIngreso: 0.0,
-        mayoIngreso: 0.0,
-        junioIngreso: 0.0,
-        julioIngreso: 0.0,
-        agostoIngreso: 0.0,
-        septiembreIngreso: 0.0,
-        octubreIngreso: 0.0,
-        noviembreIngreso: 0.0,
-        diciembreIngreso: 0.0,
-        observacion: null,
-        discapacidad: null,
-        emergenciaNombre: null,
-        emergenciaTelefono: null,
-        emergenciaTelefono2: null,
-        emergenciaDireccion: null,
-        emergenciaParentezco: null,
-        tipoSangreId: 0,
-        emergenciaAlergico: null,
-        emergenciaDiabetico: 'sT',
-        emergenciaInsodepend: 'st',
-        emergenciaPresionAlta: null,
-        emergenciaPresionBaja: null,
-        emergenciaEnTratamiento: 'st',
-        emergenciaDiagnostico: null,
-        licenciaConducir: null,
-        fechaExpiracionLicencia: new Date(Date.now()),
-        aplicaSasp: true,
-        nivelEscolar: null,
-        areaTematica: null,
-        tituloObtenido: null,
-        correoElectronico: null,
-        correoElectronico2: null,
-        recomendadoPor: null,
-      },
 
       columns: [
         { key: 'codigo', label: 'Código', _style: { width: '15%' } },
         { key: 'apellidos', label: 'Apellido', _style: { width: '15%' } },
         { key: 'nombres', label: 'Nombre', _style: { width: '15%' } },
-        { key: 'codigoIdentidad', label: 'Cédula/Pasaporte', _style: { width: '15%' } },
+        {
+          key: 'codigoIdentidad',
+          label: 'Cédula/Pasaporte',
+          _style: { width: '15%' },
+        },
         {
           key: 'programaDivision',
           label: 'Programa',
@@ -249,7 +119,6 @@ export default {
 
         { key: 'sexo', label: 'Sexo', _style: { width: '20%' } },
 
-
         {
           key: 'show_details',
           label: '',
@@ -262,7 +131,6 @@ export default {
         {
           label: 'Total Items',
           _props: {
-            color: '',
             colspan: 9,
             style: 'font-weight:bold;',
           },
@@ -274,27 +142,27 @@ export default {
           label: 'Editar',
           clickHandler: (value) => {
             this.toggleDetails(value)
-          }
+          },
         },
         {
           label: 'Eliminar',
           clickHandler: (value) => {
             this.deleteEmp(value)
-          }
+          },
         },
         {
           label: 'Evaluación',
           clickHandler: (value) => {
             this.showAccionPersonal = true
             this.selectedEmployee = value
-          }
+          },
         },
         {
           label: 'Eventualidad',
           clickHandler: (value) => {
             this.showTipoNovedad = true
             this.selectedEmployee = value
-          }
+          },
         },
         {
           label: 'Educación',
@@ -302,21 +170,21 @@ export default {
             this.showEducacion = true
             this.selectedEmployee = { ...item }
             this.selectedEmployee.nombres = `${item.nombres} ${item.apellidos}`
-          }
+          },
         },
         {
           label: 'Tarjeta',
           clickHandler: (item) => {
             this.selectedEmployee = item
             this.newTarjetaEmpleadoModal = true
-          }
+          },
         },
         {
           label: 'Ver Documentos',
           clickHandler: (item) => {
-            this.selectedEmployee = item;
+            this.selectedEmployee = item
             this.showModalDoc = true
-          }
+          },
         },
       ],
     }
@@ -330,17 +198,6 @@ export default {
         this.registroPersonal = response.data.data
       })
     },
-
-    // closeRegistroPersonalModal() {
-    //   this.showRegistroPersonalModal = false
-    //   this.getRegistroPersonal()
-    // }
-    // ,
-    // getAccionPersonal() {
-    //   Api.getAllAccionPersonal().then((response) => {
-    //     this.accionPersonal = response.data.data
-    //   })
-    // },
 
     closeRegistroPersonalModal() {
       this.showRegistroPersonalModal = false
@@ -385,36 +242,6 @@ export default {
       }
     },
 
-    // validateAge() {
-    //   if (this.calcularEdad(this.postEmpleado.fechaNacimiento) < 18) {
-    //     this.$swal({
-    //       title: 'Información',
-    //       text: 'La fecha de nacimiento seleccionada no cumple la mayoria de edad',
-    //       icon: 'Atención',
-    //       allowOutsideClick: false,
-    //       showCancelButton: true,
-    //       confirmButtonColor: '#3085d6',
-    //       cancelButtonColor: '#d33',
-    //       confirmButtonText: 'De acuerdo',
-    //     }).then((result) => {
-    //       this.lgDemo = false
-    //     })
-    //   }
-    // },
-
-    // calcularEdad(fechaNacimiento) {
-    //   var hoy = new Date()
-    //   var cumpleanos = new Date(fechaNacimiento)
-    //   var edad = hoy.getFullYear() - cumpleanos.getFullYear()
-    //   var mes = hoy.getMonth() - cumpleanos.getMonth()
-    //   if (mes < 0 || (mes === 0 && hoy.getDate() < cumpleanos.getDate())) {
-    //     edad--
-    //   }
-    //   return edad
-    // },
-    // openModal() {
-    //   this.lgDemo = true
-    // },
     closeContenedorModal(payload) {
       this.showModalDoc = payload
     },
@@ -426,89 +253,30 @@ export default {
       this.showTipoNovedad = close
     },
 
-    // closeModalssss(payload) {
-    //   this.showModalRepots = payload
-    // },
-
-    // changePrograma(e) {
-    //   Api.getDepartamentoByProgramaId(e.target.value).then((response) => {
-    //     this.departamentos = response.data.data
-    //   })
-    // },
-
-    // closeEducacion() {
-    //   this.showEducacion = false
-    // },
-
-    // formatDate(fechaIngreso) {
-    //   return new Date(fechaIngreso).toLocaleDateString('en-GB', {
-    //     day: '2-digit',
-    //     month: '2-digit',
-    //     year: 'numeric',
-    //   })
-    // },
-
-    // handleSubmitCustom01(event) {
-    //   const form = event.currentTarget
-    //   if (form.checkValidity() === false) {
-    //     event.preventDefault()
-    //     event.stopPropagation()
-    //   }
-    //   this.validatedCustom01 = true
-    // },
-
-    // getBadge(status) {
-    //   switch (status) {
-    //     case 'Active':
-    //       return 'success'
-    //     case 'Inactive':
-    //       return 'secondary'
-    //     case 'Pending':
-    //       return 'warning'
-    //     case 'Banned':
-    //       return 'danger'
-    //     default:
-    //       'primary'
-    //   }
-    // },
-
-    // getEmpleadoByID(item) {
-    //   this.showEducacion = true
-    //   this.employeeInfo = { ...item }
-    //   this.employeeInfo.nombres = `${item.nombres} ${item.apellidos}`
-    // },
-
     toggleDetails(item) {
-      this.showModal();
-      this.selectedEmployee = item;
-      // if (item.empleados !== 0 || item.variacion !== 0) {
-      //   this.empleadoValue = true
-      // } else {
-      //   this.empleadoValue = false
-      // }
-      // Api.getEmpleadoByID(item.id).then((response) => {
-      //   this.id = item.id
-      //   this.postEmpleado = response.data.data
-      // })
+      this.showModal()
+      this.selectedEmployee = item
     },
 
     submitForm(payload) {
       if (payload.id != null) {
-        Api.putEmpleado(payload.id, payload).then(() => {
-          this.show({
-            content: 'Registro actualizado correctamente',
-            closable: true,
+        Api.putEmpleado(payload.id, payload)
+          .then(() => {
+            this.show({
+              content: 'Registro actualizado correctamente',
+              closable: true,
+            })
+            this.closeRegistroPersonalModal()
+            setTimeout(this.getRegistroPersonal, 500)
           })
-          this.closeRegistroPersonalModal()
-          setTimeout(this.getRegistroPersonal, 500)
-        }).catch((error) => {
-          this.show({
-            content: error.response.data,
-            closable: true,
-            color: 'danger',
-            class: 'text-white',
+          .catch((error) => {
+            this.show({
+              content: error.response.data,
+              closable: true,
+              color: 'danger',
+              class: 'text-white',
+            })
           })
-        })
       } else {
         Api.postEmpleado(payload)
           .then(() => {
@@ -518,7 +286,8 @@ export default {
             })
             setTimeout(this.getRegistroPersonal, 500)
             this.closeRegistroPersonalModal()
-          }).catch((error) => {
+          })
+          .catch((error) => {
             this.show({
               content: error.response.data,
               closable: true,
@@ -526,41 +295,41 @@ export default {
               class: 'text-white',
             })
           })
-
-        // setTimeout(this.getRegistroPersonal, 500)
-
-        // this.lgDemo = true
-        // setTimeout(this.getRegistroPersonal, 500)
-        // ;(this.postEmpleado = {
-        //   id: 0,
-        //   ayuntamientoId: this.$ayuntamientoId,
-        //   nombre: null,
-        // }),
-        //   (this.validatedCustom01 = false)
-        // event.preventDefault()
-        // event.stopPropagation()
-        // setTimeout(this.getRegistroPersonal, 500)
       }
     },
 
     deleteEmp(item) {
-      Api.deleteEmpleado(item.id)
-        .then((response) => {
-          this.show({
-            content: response.data.message,
-            closable: true,
-            color: 'inherit',
-          })
-          setTimeout(this.getRegistroPersonal, 500)
-        })
-        .catch((error) => {
-          this.show({
-            content: error.response.data,
-            closable: true,
-            color: 'danger',
-            class: 'text-white',
-          })
-        })
+      Swal.fire({
+        position: 'center',
+        icon: 'warning',
+        title: `Estás usted seguro que quieres inactivar este empleado?`,
+        showConfirmButton: true,
+        confirmButtonText: 'Si',
+        cancelButtonText: 'No',
+        showCancelButton: true,
+        allowEscapeKey: false,
+        allowOutsideClick: false,
+      }).then((answer) => {
+        if (answer.isConfirmed) {
+          Api.deleteEmpleado(item.id)
+            .then((response) => {
+              this.show({
+                content: response.data.message,
+                closable: true,
+                color: 'inherit',
+              })
+              setTimeout(this.getRegistroPersonal, 500)
+            })
+            .catch((error) => {
+              this.show({
+                content: error.response.data,
+                closable: true,
+                color: 'danger',
+                class: 'text-white',
+              })
+            })
+        }
+      })
     },
   },
 
@@ -579,7 +348,7 @@ export default {
   watch: {
     showRegistroPersonalModal() {
       this.getRegistroPersonal()
-    }
+    },
   },
 }
 </script>
